+++ test-support/helpers/fill-in-label.js
@@ -1,25 +1,45 @@
 import Ember from 'ember';
 
+const { run } = Ember;
+
 export default function(app, i18n, value) {
   const {
     clickOn,
+    find,
+    findWithAssert,
     fillIn,
-    t
+    t,
+    wait
   } = app.testHelpers;
   const text = t(i18n);
   const label = `label:contains("${text}")`;
-  const selector = [
-    `${label} input:not([type="checkbox"])`,
-    `${label} + input:not([type="checkbox"])`,
-    `${label} textarea`,
-    `${label} + textarea`,
-    `${label} select`,
-    `${label} + select`,
-  ].join(',');
 
-  return fillIn(selector, value)
-    .then(
-      Ember.K,
-      () => clickOn(i18n)
-    );
+  run(() => {
+    const $label = find(label);
+    const id = $label.prop('for');
+
+    if (id) {
+      const $element = findWithAssert(`#${id}`);
+
+      if ($element.prop('type') === 'checkbox') {
+        $element.prop('checked', value);
+      } else {
+        $element.val(value);
+      }
+    } else {
+      const nested = [
+        `${label} input:not([type="checkbox"])`,
+        `${label} textarea`,
+        `${label} select`,
+      ].join(',');
+
+      fillIn(nested, value)
+        .then(
+          () => wait(),
+          () => clickOn(i18n)
+        );
+    }
+
+    return wait();
+  });
 };